fib.c
As the current value is sum of previous two calculations only. This indicates that we don't have to store all the previous values. Instead if we store just last two values,we can calculate the current value.

Time Complexity: O(n)

Space Complexity: O(1)

Note : This method may not be applicable for all problems